generify upload form
